% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpi.R
\name{sim_pi}
\alias{sim_pi}
\title{Reproduce memory leak error by simulating pi in parallel}
\usage{
sim_pi(n, cl, size = 1000)
}
\arguments{
\item{n}{Integer, number of simulations.}

\item{cl}{an object of class \code{cluster} (see \link[parallel:makeCluster]{parallel::makeCluster}).}

\item{size}{Integer. Size of each simulation.}
}
\description{
Reproduce memory leak error by simulating pi in parallel
}
\examples{
# PSOCK is OK
cl <- makePSOCKcluster(2)
ans0 <- sim_pi(100, cl)
stopCluster(cl)

# Fork is not
if (.Platform$OS.type == "unix") {
  cl <- makeForkCluster(2)
  ans0 <- sim_pi(100, cl)
  stopCluster(cl) # Produces errors catched by valgrind
}
}
